[merge.c, merge_mono.c] my_cost now returns a double, which avoids overflow 	  problems on 32-bit machines. Moreover, the double precision should 	  be largely enough.
	  I did test this patch on revision 1717 and it produced the correct
	  final matrix on the c59, unfortunately I could not test it with
	  revision 1719 since that revision was not functional...


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1720 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/merge.c
@@ -201,10 +201,9 @@ main(int argc, char *argv[])
 #endif
 
     gzip_close(rep.outfile, outname);
-    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%lu\n",
+    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%1.0f\n",
 	    mat.rem_nrows, mat.rem_ncols, mat.rem_nrows-mat.rem_ncols,
-	    mat.weight,
-	    ((unsigned long)mat.rem_nrows) * mat.weight);
+	    mat.weight, (double) mat.rem_nrows * (double) mat.weight);
 #if TEX
     fprintf(stderr, "\\end{verbatim}\n");
 #endif
